+++ packages/ice/src/utils/generateEntry.ts
@@ -1,6 +1,8 @@
 import * as path from 'path';
 import fse from 'fs-extra';
-import type { ServerContext, RenderMode, AppConfig } from '@ice/runtime';
+import consola from 'consola';
+import type { ServerContext, RenderMode, AppConfig, DistType } from '@ice/runtime';
+import type { UserConfig } from '../types/userConfig.js';
 import getRoutePaths from './getRoutePaths.js';
 import dynamicImport from './dynamicImport.js';
 import { logger } from './logger.js';
@@ -13,9 +15,14 @@ interface Options {
   documentOnly: boolean;
   routeType: AppConfig['router']['type'];
   renderMode?: RenderMode;
+  distType: UserConfig['output']['distType'];
 }
 
-export default async function generateHTML(options: Options) {
+interface EntryResult {
+  outputPaths: Array<string>;
+}
+
+export default async function generateEntry(options: Options): Promise<EntryResult> {
   const {
     rootDir,
     entry,
@@ -24,6 +31,8 @@ export default async function generateHTML(options: Options) {
     renderMode,
     routeType,
   } = options;
+
+  const distType = typeof options.distType === 'string' ? [options.distType] : options.distType;
 
   let serverEntry;
   try {
@@ -37,14 +46,36 @@ export default async function generateHTML(options: Options) {
   const routes = getRouterManifest(rootDir);
   // When enable hash-router, only generate one html(index.html).
   const paths = routeType === 'hash' ? ['/'] : getRoutePaths(routes);
+  const outputPaths = [];
   for (let i = 0, n = paths.length; i < n; i++) {
     const routePath = paths[i];
-    const htmlContent = await renderHTML({ routePath, serverEntry, documentOnly, renderMode });
-    await writeFile(
-      await generateHTMLPath({ rootDir, routePath, outputDir }),
-      htmlContent,
-    );
+    const {
+      htmlOutput,
+      jsOutput,
+    } = await renderEntry({ routePath, serverEntry, documentOnly, renderMode, distType });
+
+    if (htmlOutput) {
+      const path = await generateHTMLPath({ rootDir, routePath, outputDir });
+      await writeFile(
+        path,
+        htmlOutput,
+      );
+      outputPaths.push(path);
+    }
+
+    if (jsOutput) {
+      const path = await generateJSPath({ rootDir, routePath, outputDir });
+      await writeFile(
+        path,
+        jsOutput,
+      );
+      outputPaths.push(path);
+    }
   }
+
+  return {
+    outputPaths,
+  };
 }
 
 const writeFile = async (file: string, content: string) => {
@@ -72,16 +103,38 @@ async function generateHTMLPath(
   return path.join(outputDir, fileName);
 }
 
-async function renderHTML(
+async function generateJSPath(
+  {
+    rootDir,
+    routePath,
+    outputDir,
+  }: {
+    rootDir: string;
+    routePath: string;
+    outputDir: string;
+  },
+) {
+  // Win32 do not support file name with
+  const fileName = routePath === '/' ? 'index.js' : `${routePath.replace(/\/:/g, '/$')}.js`;
+  if (fse.existsSync(path.join(rootDir, 'public', fileName))) {
+    consola.warn(`${fileName} is overwrite by framework, rename file name if it is necessary.`);
+  }
+
+  return path.join(outputDir, fileName);
+}
+
+async function renderEntry(
   {
     routePath,
     serverEntry,
     documentOnly,
+    distType = ['html'],
     renderMode,
   }: {
     routePath: string;
     serverEntry: any;
     documentOnly: boolean;
+    distType?: DistType;
     renderMode?: RenderMode;
   },
 ) {
@@ -90,11 +143,23 @@ async function renderHTML(
       url: routePath,
     } as any,
   };
-  const { value: html } = await serverEntry.renderToHTML(serverContext, {
+
+  // renderToEntry exported when disType includes javascript .
+  const render = distType.includes('javascript') ? serverEntry.renderToEntry : serverEntry.renderToHTML;
+
+  const {
+    value,
+    jsOutput,
+  } = await render(serverContext, {
     renderMode,
     documentOnly,
     routePath,
     serverOnlyBasename: '/',
+    distType,
   });
-  return html;
+
+  return {
+    htmlOutput: value,
+    jsOutput,
+  };
 }