+++ packages/rax-compat/src/runtime/jsx-dev-runtime.ts
@@ -1,15 +1,7 @@
 
-import { jsxs, jsx } from './impl.js';
-export { Fragment } from './impl.js';
-export function jsxDEV(
-  type: any,
-  props: object,
-  key: string,
-  isStaticChildren: boolean,
-  source: object,
-  self: any,
-) {
-  return isStaticChildren
-    ? jsxs(type, props, key, source, self)
-    : jsx(type, props, key, source, self);
-}
+// @ts-ignore
+import { jsxDEV as _jsxDEV } from 'react/jsx-dev-runtime';
+import { createJSXElementFactory } from '../compat/element';
+
+export { Fragment } from 'react';
+export const jsxDEV = createJSXElementFactory(_jsxDEV);